+++ src/integration/extradelight/java/plus/dragons/createcentralkitchen/integration/extradelight/ExtraDelightIntegration.java
@@ -20,6 +20,10 @@ package plus.dragons.createcentralkitchen.integration.extradelight;
 
 import com.lance5057.extradelight.ExtraDelightBlockEntities;
 import com.simibubi.create.api.packager.unpacking.UnpackingHandler;
+import java.util.*;
+import net.minecraft.server.packs.resources.ResourceManagerReloadListener;
+import net.minecraft.world.item.crafting.Recipe;
+import net.minecraft.world.item.crafting.RecipeHolder;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.bus.api.SubscribeEvent;
@@ -28,6 +32,7 @@ import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.neoforged.fml.event.lifecycle.FMLConstructModEvent;
 import net.neoforged.fml.loading.FMLLoader;
 import net.neoforged.neoforge.common.NeoForge;
+import net.neoforged.neoforge.event.AddReloadListenerEvent;
 import plus.dragons.createcentralkitchen.common.CCKCommon;
 import plus.dragons.createcentralkitchen.integration.ModIntegration;
 import plus.dragons.createcentralkitchen.integration.extradelight.packager.ChillerUnpackingHandler;
@@ -47,11 +52,15 @@ public class ExtraDelightIntegration {
     }
 
     public static class Common {
+        public static final Set<Collection<RecipeHolder<? extends Recipe<?>>>> RELOADABLE_RECIPES = new HashSet<>();
+        private static final ResourceManagerReloadListener RELOAD_LISTENER = resourceManager -> RELOADABLE_RECIPES.forEach(Collection::clear);
+
         @SubscribeEvent
         public void construct(final FMLConstructModEvent event) {
             ExtraDelightArmInteractionPointTypes.register();
             // Extra Delight itself makes every mixing recipe to create-mixing recipe, so we don't need to do conversion for that.
             NeoForge.EVENT_BUS.register(ExtraDelightRecipeConverters.class);
+            NeoForge.EVENT_BUS.addListener(Common::addReloadListeners);
         }
 
         @SubscribeEvent
@@ -66,6 +75,10 @@ public class ExtraDelightIntegration {
             ExtraDelightBlockEntities.CHILLER.get()
                     .getValidBlocks()
                     .forEach(block -> UnpackingHandler.REGISTRY.register(block, new ChillerUnpackingHandler()));
+        }
+
+        public static void addReloadListeners(AddReloadListenerEvent event) {
+            event.addListener(RELOAD_LISTENER);
         }
     }