+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-## Create: Central Kitchen 2.2.2b
+## Create: Central Kitchen 2.2.3
 
 ### Fix
-* Fix extreme tick rate spike when Extra Delight mixer/press compat is working
-* Fix a registration issue when Json Things and Hostile Neural Networks are present.
+* Fix Some of Extra Delight Compat Recipes are not reloadable
+
+### Translation
+* Add PT_BR localization by PrincessStellar